requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
asgi-correlation-id==4.3.4
asn1crypto==1.5.1
attrs==24.3.0
black==24.10.0
build==1.2.2.post1
cachetools==5.5.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cloud-sql-python-connector==1.15.0
cryptography==44.0.0
dataclasses-json==0.6.7
distro==1.9.0
dnspython==2.7.0
docker==7.1.0
fastapi==0.115.6
flake8==7.1.1
Flake8-pyproject==1.2.3
frozenlist==1.5.0
google-api-core==2.24.0
google-auth==2.37.0
google-cloud-core==2.4.1
google-cloud-secret-manager==2.22.0
google-cloud-storage==2.19.0
google-crc32c==1.6.0
google-resumable-media==2.7.2
google_search_results==2.4.2
googleapis-common-protos==1.66.0
grpc-google-iam-v1==0.14.0
grpcio==1.69.0
grpcio-status==1.69.0
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
httpx-sse==0.4.1
idna==3.10
importlib-resources==5.13.0
iniconfig==2.0.0
jiter==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.26
langchain-community==0.3.27
langchain-core==0.3.68
langchain-openai==0.3.27
langchain-text-splitters==0.3.8
langgraph==0.5.2
langgraph-checkpoint==2.1.0
langgraph-prebuilt==0.5.2
langgraph-sdk==0.1.72
langsmith==0.4.5
logging-http-client==2.32.3.9
markdown-it-py==3.0.0
marshmallow==3.26.1
mccabe==0.7.0
mdurl==0.1.2
multidict==6.1.0
mypy-extensions==1.0.0
numpy==2.3.1
openai==1.93.3
orjson==3.10.18
ormsgpack==1.10.0
packaging==24.2
pathspec==0.12.1
pg8000==1.31.2
pip-tools==7.4.1
platformdirs==4.3.6
pluggy==1.5.0
prometheus-fastapi-instrumentator==7.0.0
prometheus_client==0.21.1
propcache==0.2.1
proto-plus==1.25.0
protobuf==5.29.2
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pyclean==3.0.0
pycodestyle==2.12.1
pycparser==2.22
pydantic==2.10.4
pydantic-settings==2.10.1
pydantic-xml==2.14.1
pydantic_core==2.27.2
pyflakes==3.2.0
Pygments==2.19.1
pyproject_hooks==1.2.0
pytest==8.3.4
pytest-env==1.1.5
pytest-mock==3.14.0
pytest-order==1.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
rsa==4.9
scramp==1.4.5
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.3
structlog==24.4.0
tenacity==9.1.2
testcontainers==4.9.0
tiktoken==0.9.0
tqdm==4.67.1
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.32.1
wheel==0.45.1
wiremock==2.6.1
wrapt==1.17.0
xmltodict==0.14.2
xxhash==3.5.0
yarl==1.18.3
zstandard==0.23.0
